Deprecate usage of internal registries completely
Internal registries are supposed for internal use only. They change the protocols and signatures all the time. Backward compatibility will not be ensured in the future.

The only usable registry is `OperatorRegistry` or `SmartOperatorRegistry`, and only to the extent that it is created and used in the `registry=` kwarg to `@kopf.on...` decorators and `operator()` coroutine and related functions. All public fields and methods of these classes, such as selecting the handlers or fields, are also deprecated.